+++ tests/Benchmark/BaseBenchCase.php
@@ -3,7 +3,7 @@
 namespace Phpactor\Tests\Benchmark;
 
 use Phpactor\Tests\IntegrationTestCase;
-use Phpactor\Console\Application;
+use Phpactor\Application;
 use Symfony\Component\Console\Output\BufferedOutput;
 use Symfony\Component\Console\Input\ArrayInput;
